feat: integrate CHANGELOG.md CI in Main Merge
We require that CHANGELOG.md has been updated with new Sem Ver Release entry before enabling Auto Merge to PR to main branch. If CHANGELOG.md is NOT updated according to the Rules, then the 'go-auto-merge-main.yml' Workflow will fail. Also, we add a fix on condition logic when the 'auto-deploy' label needs to be registered in the Repository.