feat: allow User to singal automatic CHANGELOG.md update before 'Release Me' Phase 2!
Allow User to see the content that was automatically generated for new Release Changelog.md content,
on a PR to main comment. Also, allow user to turn on/off automatic changelog.md update using a
payload posted on another PR Comment! We also make the 'Auto Merg to Main' Workflow require a
CHnagelog MD Update or it fails. It also tried to do the automatica update if changelog was not
touched at all in incoming beanch nad if the payload has the bool value set to true. Payload is
posted as we said, as a comment in the Pr to main.